INT-4390: Phase I: flow context/registration cycle
JIRA: https://jira.spring.io/browse/INT-4390

- extract interfaces
- rename implementations to `Standard...`
